Move 8 of 9 template filters and globals into a blueprint
